chore: implement header component
